Rename oauth route to google-auth
